+++ blueno/pipeline/__init__.py
@@ -1,2 +1,3 @@
 from .bluenot import start_train
 from .bluenop import start_preprocess
+from .bluenov import start_validation, get_validation_params
